requirements.txt
tensorflow==2.10.0
Flask
Pillow
numpy==1.23.5
scipy